📑 :: 습관성공 엔티티 수정
